# internal/handlers/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "ShieldNet 360 Access Platform API",
    "description": "HTTP surface for the ZTNA access platform (see docs/architecture.md \u00a72). Covers access requests, reviews, grants, policy drafting / simulation, AI explain / suggest, and inbound SCIM (auto-sync users).\n\nThis spec is regenerated by `scripts/generate-swagger.sh` whenever a handler-level change ships (see CONTRIBUTING.md).",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/",
      "description": "Same-origin server. The access API is served by cmd/ztna-api."
    }
  ],
  "paths": {
...
    "/access/connectors": {
      "get": {
        "summary": "List access connectors with last-sync timestamps and capability flags",
        "tags": [
          "access-connectors"
        ],
        "parameters": [
          {
            "name": "workspace_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "workspace_id required"
          }
        }
      },
      "post": {
        "summary": "Connect a new access connector \u2014 validates, probes, encrypts, persists, and enqueues an initial sync_identities job",
        "tags": [
          "access-connectors"
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Validation failed / unknown provider"
          },
          "409": {
            "description": "Duplicate connector for (workspace, provider, connector_type)"
          }
        }
      }
    },
    "/access/connectors/{id}": {
      "delete": {
        "summary": "Soft-delete an access connector \u2014 revokes active grants and deregisters SSO",
        "tags": [
          "access-connectors"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Connector not found"
          }
        }
      }
    },
...
    "/access/connectors/{id}/secret": {
      "put": {
        "summary": "Rotate connector credentials \u2014 re-validates, re-encrypts, and atomically updates the secrets column",
        "tags": [
          "access-connectors"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Validation failed"
          },
          "404": {
            "description": "Connector not found"
          }
        }
      }
    },
    "/access/connectors/{id}/sync": {
      "post": {
        "summary": "Trigger an immediate sync_identities job for the supplied connector",
        "tags": [
          "access-connectors"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Job enqueued"
          },
          "404": {
            "description": "Connector not found"
          }
        }
      }
    },
...
    "/access/connectors/batch-status": {
      "post": {
        "summary": "Batch connector health lookup \u2014 avoids N+1 calls from the Admin UI connector list page",
        "description": "Accepts an array of connector IDs and returns each connector's health view in a single response. Per-entry errors (e.g. connector deleted) are inlined rather than aborting the whole request.",
        "tags": [
          "access-connectors"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchStatusRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchStatusResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error (empty array, oversized, malformed JSON)"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConnectorHealth": {
        "type": "object",
        "description": "Operational health view of a single access connector.",
        "properties": {
          "connector_id": {
            "type": "string"
          },
          "provider": {
            "type": "string"
          },
          "connector_type": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "credential_expired_time": {
            "type": "string",
            "format": "date-time"
          },
          "last_sync_times": {
            "type": "object",
            "additionalProperties": {
              "type": "string",
              "format": "date-time"
            },
            "description": "Map of sync kind (identity/group/audit) to last-sync timestamp"
          },
          "stale_audit": {
            "type": "boolean"
          }
        },
        "required": [
          "connector_id",
          "provider",
          "status",
          "stale_audit"
        ]
      },
      "BatchStatusRequest": {
        "type": "object",
        "description": "POST body for /access/connectors/batch-status.",
        "properties": {
          "connector_ids": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "maxItems": 200,
            "description": "Array of connector IDs to look up"
          }
        },
        "required": [
          "connector_ids"
        ]
      },
      "BatchStatusResponse": {
        "type": "object",
        "description": "Response envelope for /access/connectors/batch-status. Entries preserve the caller order (after dedup).",
        "properties": {
          "entries": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BatchStatusEntry"
            }
          }
        },
        "required": [
          "entries"
        ]
      },
      "BatchStatusEntry": {
        "type": "object",
        "description": "Per-connector result. Exactly one of health, not_found, or error is populated.",
        "properties": {
          "connector_id": {
            "type": "string"
          },
          "not_found": {
            "type": "boolean"
          },
          "health": {
            "$ref": "#/components/schemas/ConnectorHealth"
          },
          "error": {
            "type": "string"
          }
        },
        "required": [
          "connector_id"
        ]
      },
...
    }
  }
}